Make Thumb use pointer events
