--- src/Uno.UI/UI/Xaml/Controls/Primitives/Thumb.net.cs
@@ -1,25 +1,0 @@
-using System;
-using Uno.UI;
-using Windows.Foundation;
-using Windows.Foundation.Metadata;
-
-namespace Windows.UI.Xaml.Controls.Primitives
-{
-	public sealed partial class Thumb : Control
-	{		
-		internal void StartDrag(Point location)
-		{
-			throw new NotImplementedException();
-		}
-
-		internal void CompleteDrag(Point location)
-		{
-			throw new NotImplementedException();
-		}
-
-		internal void DeltaDrag(Point location)
-		{
-			throw new NotImplementedException();
-		}
-	}
-}
